refactor: cleanup Stripe logic
