+++ deno.json
@@ -32,7 +32,7 @@
     "twind-preset-tailwind/": "https://esm.sh/@twind/preset-tailwind@1.1.4/",
     "twind-preset-ext": "https://esm.sh/@twind/preset-ext@1.0.7/",
     "std/": "https://raw.githubusercontent.com/lino-levan/deno_std/feat-ulid/",
-    "stripe": "./stripe.ts",
+    "stripe": "npm:/stripe@12.6.0",
     "feed": "https://esm.sh/feed@4.2.2",
     "kv_oauth": "https://deno.land/x/deno_kv_oauth@v0.6.1/mod.ts",
     "tabler_icons_tsx/": "https://deno.land/x/tabler_icons_tsx@0.0.4/tsx/",
