+++ routes/account/manage.ts
@@ -1,6 +1,6 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import type { RouteContext } from "$fresh/server.ts";
-import { stripe } from "@/utils/payments.ts";
+import { stripe } from "@/utils/stripe.ts";
 import type { SignedInState } from "@/middleware/session.ts";
 import { redirect } from "@/utils/http.ts";
 import { errors } from "std/http/http_errors.ts";
