+++ routes/account/upgrade.ts
@@ -1,6 +1,6 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import type { RouteContext } from "$fresh/server.ts";
-import { stripe } from "@/utils/payments.ts";
+import { isStripeEnabled, stripe } from "@/utils/stripe.ts";
 import type { SignedInState } from "@/middleware/session.ts";
 import { redirect } from "@/utils/http.ts";
 import { errors } from "std/http/http_errors.ts";
@@ -13,7 +13,7 @@ export default async function AccountUpgradePage(
   _req: Request,
   ctx: RouteContext<undefined, SignedInState>,
 ) {
-  if (stripe === undefined) throw new errors.NotFound();
+  if (!isStripeEnabled()) throw new errors.NotFound();
   if (STRIPE_PREMIUM_PLAN_PRICE_ID === undefined) {
     throw new Error(
       '"STRIPE_PREMIUM_PLAN_PRICE_ID" environment variable not set',
